+++ src/containers/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import '../styles/App.css';
 import { BrowserRouter, Route } from 'react-router-dom'
-import HomePage from './HomePage'
+import HomePage from '../components/HomePage'
 
 class App extends Component {
   constructor(props) {
@@ -21,6 +21,7 @@ class App extends Component {
         <main>
           <BrowserRouter>
             <Route path="/" component={ HomePage }/>
+            
           </BrowserRouter>
         </main>
       </div>
