+++ src/containers/WineList.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import '../styles/WineList.css';
-import WineCard from './WineCard'
+import WineCard from '../components/WineCard'
 
 class WineList extends Component {
   constructor(props) {
