+++ src/styles/App.css
@@ -20,6 +20,6 @@ font-family: 'Tangerine', cursive;*/
 }
 
 .App-glass {
-  width: 8%;
+  width: 10%;
   height: 40%;
 }
