+++ src/styles/HomePage.css
@@ -2,7 +2,8 @@
 
 .Home-page-container {
   display: flex;
-  flex-direction: row;
+  flex-flow: row;
+  justify-content: space-around;
 }
 
 .Home-page {
